requirements.txt
fastapi==0.74.0
python-dotenv==0.19.2
uvicorn==0.17.5
sqlalchemy==1.4.44
psycopg2==2.9.5
